Change API variable types to be cpp stock types

+++ src/granular/GranularDefines.h
@@ -17,9 +17,9 @@ typedef uint16_t subVoxelPos_t;
 typedef uint64_t voxelID_t;
 typedef int oriQ_t;  // orientation quaternion data type
 typedef unsigned int bodyID_t;
-typedef unsigned int materialsOffset_t;
-typedef unsigned int clumpBodyInertiaOffset_t;
-typedef unsigned int clumpComponentOffset_t;
+typedef unsigned short int materialsOffset_t;
+typedef unsigned short int clumpBodyInertiaOffset_t;
+typedef unsigned short int clumpComponentOffset_t;
 
 // typedef unsigned int stateVectors_default_t; // what's this for??
 // typedef unsigned int distinctSphereRelativePositions_default_t;
